cloud_functions/ocr_extraction/requirements.txt
functions-framework==3.*
google-cloud-storage==2.10.0
google-cloud-bigquery==3.11.4
google-cloud-pubsub==2.18.4
anthropic==0.18.1
pdf2image==1.16.3
Pillow==10.2.0

